pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ortflix-telegram-bot"
description = "Telegram bot for Overseerr media requests"
dynamic = ["version"]
readme = "README.md"
license = "MIT"
authors = [
    {name = "okdesign21"}
]
keywords = ["telegram", "bot", "overseerr", "ortflix", "media"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Communications :: Chat",
    "Topic :: Home Automation",
]
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=22.6",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "aiohttp>=3.13.3",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[project.urls]
Homepage = "https://github.com/okdesign21/ortflix-telegram"
Repository = "https://github.com/okdesign21/ortflix-telegram"
Documentation = "https://github.com/okdesign21/ortflix-telegram"
Issues = "https://github.com/okdesign21/ortflix-telegram/issues"

[project.scripts]
ortflix-telegram-bot = "bot:main"

[tool.setuptools]
py-modules = ["bot"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "C", "B"]

[tool.ruff.lint.isort]
known-first-party = ["bot"]

[tool.setuptools_scm]
# Version derived from git tags (e.g., v0.1.0 -> 0.1.0)
write_to = "_version.py"
local_scheme = "no-local-version"
fallback_version = "0.1.0"
